# project/public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/hospitals": {
      "post": {
        "description": "Create a new hospital organization (ADMIN or HOSPITAL role)",
        "operationId": "HospitalController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHospitalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Hospital created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create hospital",
        "tags": [
          "Hospitals"
        ]
      },
      "get": {
        "description": "Get list of all hospitals with optional search",
        "operationId": "HospitalController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query for hospital name or location",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hospitals retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all hospitals",
        "tags": [
          "Hospitals"
        ]
      }
    },
    "/hospitals/{id}": {
      "get": {
        "description": "Get detailed information about a specific hospital",
        "operationId": "HospitalController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Hospital ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hospital retrieved successfully"
          },
          "404": {
            "description": "Hospital not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get hospital by ID",
        "tags": [
          "Hospitals"
        ]
      },
      "put": {
        "description": "Update hospital information",
        "operationId": "HospitalController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Hospital ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHospitalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hospital updated successfully"
          },
          "404": {
            "description": "Hospital not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update hospital",
        "tags": [
          "Hospitals"
        ]
      },
      "delete": {
        "description": "Delete a hospital (ADMIN only)",
        "operationId": "HospitalController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Hospital ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hospital deleted successfully"
          },
          "404": {
            "description": "Hospital not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete hospital",
        "tags": [
          "Hospitals"
        ]
      }
    },
    "/hospitals/{id}/capacity": {
      "put": {
        "description": "Update bed capacity information (HOSPITAL role)",
        "operationId": "HospitalController_updateCapacity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Hospital ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHospitalCapacityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Capacity updated successfully"
          },
          "404": {
            "description": "Hospital not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update hospital capacity",
        "tags": [
          "Hospitals"
        ]
      }
    },
    "/hospitals/{id}/accept-emergency": {
      "post": {
        "description": "Hospital accepts an emergency request",
        "operationId": "HospitalController_acceptEmergency",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Hospital ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptEmergencyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Emergency accepted successfully"
          },
          "404": {
            "description": "Hospital or emergency not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accept emergency request",
        "tags": [
          "Hospitals"
        ]
      }
    },
    "/hospitals/nearby/{latitude}/{longitude}": {
      "get": {
        "description": "Find hospitals within a specified radius from coordinates",
        "operationId": "HospitalController_findNearbyHospitals",
        "parameters": [
          {
            "name": "latitude",
            "required": true,
            "in": "path",
            "description": "Latitude coordinate",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "longitude",
            "required": true,
            "in": "path",
            "description": "Longitude coordinate",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "radius",
            "required": false,
            "in": "query",
            "description": "Search radius in kilometers",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Nearby hospitals retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Find nearby hospitals",
        "tags": [
          "Hospitals"
        ]
      }
    },
    "/hospitals/emergency-responses/{id}": {
      "put": {
        "description": "Update status of an emergency response",
        "operationId": "HospitalController_updateEmergencyResponseStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Emergency response ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Emergency response status updated"
          },
          "404": {
            "description": "Emergency response not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update emergency response status",
        "tags": [
          "Hospitals"
        ]
      },
      "get": {
        "description": "Retrieve details of an emergency response",
        "operationId": "HospitalController_getEmergencyResponse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Emergency response ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Emergency response retrieved"
          },
          "404": {
            "description": "Emergency response not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get emergency response",
        "tags": [
          "Hospitals"
        ]
      }
    },
    "/hospitals/emergency-responses/{id}/notify-rescue": {
      "post": {
        "description": "Notify rescue team about an emergency response",
        "operationId": "HospitalController_notifyRescueTeam",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Emergency response ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rescue team notified"
          },
          "404": {
            "description": "Emergency response not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Notify rescue team",
        "tags": [
          "Hospitals"
        ]
      }
    },
    "/hospitals/emergency-responses/{id}/status": {
      "patch": {
        "description": "Patch status of an emergency response",
        "operationId": "HospitalController_updateEmergencyResponseStatusManual",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Emergency response ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "New status",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Emergency response status updated"
          },
          "404": {
            "description": "Emergency response not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Manually update emergency response status",
        "tags": [
          "Hospitals"
        ]
      }
    },
    "/hospitals/{id}/reports": {
      "get": {
        "description": "Get reports for a specific hospital",
        "operationId": "HospitalController_getReports",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Hospital ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reports retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get hospital reports",
        "tags": [
          "Hospitals"
        ]
      }
    },
    "/hospitals/{id}/stats": {
      "get": {
        "description": "Get statistical data for a hospital",
        "operationId": "HospitalController_getStats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Hospital ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": false,
            "in": "query",
            "description": "Time period (day, week, month, year)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date (ISO format)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date (ISO format)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get hospital statistics",
        "tags": [
          "Hospitals"
        ]
      }
    },
...
    "/hospitals/{id}/reports/generate": {
      "post": {
        "description": "Generate a new report for the hospital",
        "operationId": "HospitalController_generateReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Hospital ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report generated successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate hospital report",
        "tags": [
          "Hospitals"
        ]
      }
    },
...
  },
  "info": {
    "title": "Emergency Project API",
    "description": "\n      ## Emergency Response Management System API\n      \n      This API provides comprehensive endpoints for managing emergency requests, hospitals, rescue teams, and related services.\n      \n      ### Features:\n      - **Authentication**: OAuth 2.0 (Google, Facebook, Apple) and Email/Password\n      - **Emergency Requests**: Create and manage SOS emergency requests\n      - **Hospital Management**: Hospital capacity, emergency acceptance\n      - **Rescue Teams**: Team management and assignment\n      - **Dashboard**: Real-time statistics and monitoring\n      - **Notifications**: Real-time push notifications\n      - **Reports**: Generate and download reports\n      - **User Settings**: Manage user preferences and settings\n      \n      ### Authentication:\n      Most endpoints require JWT Bearer token authentication. Use the /auth endpoints to obtain tokens.\n    ",
    "version": "1.0",
    "contact": {
      "name": "Emergency Project Team",
      "url": "https://emergency-project.com",
      "email": "support@emergency-project.com"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.emergency-project.com",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateHospitalDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateHospitalDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateHospitalCapacityDto": {
        "type": "object",
        "properties": {}
      },
      "AcceptEmergencyDto": {
        "type": "object",
        "properties": {}
      },
      "GenerateReportDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}